Merge 'unknown' and 'linux' platform identifiers in stats
The dw.fc_usage table contains platform entries for
'unknown' and 'linux'. The 'unknown' identifier is a legacy
identifier used in earlier version of the browser.

This commit creates a new materialized view that merges
the two identifiers into a single value 'linux'

This is the first in a series of commits that will be
required to completely merge the identifiers within the
reporting system.

Changes

  * Create a new materialized view, merging the unknown and linux
    identifiers.
  * Add materialized view to list for updating on a set schedule.
  * Modify API to use new materialized view

Auditors: @bbondy